+++ models/src/main.js
@@ -5,9 +5,11 @@ export type * from './tokens';
 export type * from './user';
 export type * from './api';
 export type * from './access';
+export type * from './message';
 */
 
 export * from './tokens.js';
 export * from './user.js';
 export * from './api.js';
 export * from './access.js';
+export * from './message.js';